Move re-usable request handlers to paths-handlers folder.
