+++ api/src/paths/dwc/view-occurrences.test.ts
@@ -2,10 +2,11 @@ import chai, { expect } from 'chai';
 import { describe } from 'mocha';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
-import * as view_occurrences from './view-occurrences';
-import * as db from '../../database/db';
-import * as occurrence_view_queries from '../../queries/occurrence/occurrence-view-queries';
 import SQL from 'sql-template-strings';
+import * as db from '../../database/db';
+import { CustomError } from '../../errors/CustomError';
+import * as occurrence_view_queries from '../../queries/occurrence/occurrence-view-queries';
+import * as view_occurrences from './view-occurrences';
 
 chai.use(sinonChai);
 
@@ -68,8 +69,10 @@ describe('getOccurrencesForView', () => {
       await result(sampleReq, (null as unknown) as any, (null as unknown) as any);
       expect.fail();
     } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Missing required request body param `occurrence_submission_id`');
+      expect((actualError as CustomError).status).to.equal(400);
+      expect((actualError as CustomError).message).to.equal(
+        'Missing required request body param `occurrence_submission_id`'
+      );
     }
   });
 
@@ -93,8 +96,8 @@ describe('getOccurrencesForView', () => {
       );
       expect.fail();
     } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Failed to build SQL get occurrences for view statement');
+      expect((actualError as CustomError).status).to.equal(400);
+      expect((actualError as CustomError).message).to.equal('Failed to build SQL get occurrences for view statement');
     }
   });
 
@@ -125,8 +128,8 @@ describe('getOccurrencesForView', () => {
       );
       expect.fail();
     } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Failed to get occurrences view data');
+      expect((actualError as CustomError).status).to.equal(400);
+      expect((actualError as CustomError).message).to.equal('Failed to get occurrences view data');
     }
   });
